requirements.txt
Django==4.1
djangorestframework==3.14.0
psycopg2==2.9.6
requests==2.29.0
Pillow==9.5.0
celery[redis]==5.2.7
python-dotenv==1.0.0
gunicorn==20.1.0